Refactor: Migrate Backups tab to Nuxt UI
- Rename Backups.vue → BackupsTab.vue for naming consistency
- Wrap in BaseTab, use UiBox for content sections
- Replace custom modal with Dialog component
- Replace anchor buttons with UButton (icons, variants)
- Replace raw inputs with UInput/UTextarea
- Move backup action to fixed bottom toolbar
- Remove all scoped CSS in favour of Tailwind utilities
