+++ src/js/vue_tab_registry.js
@@ -5,7 +5,7 @@ import PortsTab from "../components/tabs/PortsTab.vue";
 import ServosTab from "../components/tabs/ServosTab.vue";
 import ConfigurationTab from "../components/tabs/ConfigurationTab.vue";
 import UserProfileTab from "../components/tabs/UserProfile.vue";
-import BackupsTab from "../components/tabs/Backups.vue";
+import BackupsTab from "../components/tabs/BackupsTab.vue";
 import LoggingTab from "../components/tabs/LoggingTab.vue";
 import GpsTab from "../components/tabs/GpsTab.vue";
 import AuxiliaryTab from "../components/tabs/AuxiliaryTab.vue";
